Overhauled config and script loading system
Scripts are now dynamically loaded from files in the scripts subpackage using
importlib and inspect. The bot engine has zero knowledge of individual scripts
beyond interfacing attributes and methods. The scripts themselves are also
slightly modified to adapt to the new system.

The default configurations hardcoded inside the bot engine are removed, because
no matter what a config file is required. The removal of default configurations
just means that a few more settings in the config file are mandatory.